feat(restendpoint): add rest endpoint to add site monitor

+++ db/knex_migrations/2025-06-15-0001-manual-monitor-fix.js
@@ -2,7 +2,7 @@
 exports.up = function (knex) {
     return knex.schema
         .alterTable("monitor", function (table) {
-            table.smallint("manual_status").alter();
+            table.integer("manual_status").alter();
         });
 };
